+++ packages/booking-interface/utils.js
@@ -13,4 +13,31 @@ const adress = new RegExp(
 
 const extractAdresses = (str) => adress.exec(str)
 
-module.exports = { scanAddress, extractAdresses, adress }
+const formatRegexResult = (res) =>
+  res
+    .map((res) => res.groups)
+    .filter((group) => group.street && group.nr && group.zipcode)
+    .map((x) =>
+      // do something better here
+      Object.assign({}, x, { zipcode: x.zipcode.replace('SE-', '') })
+    )
+
+const peliasResToSuggestedRecipient = (r) => ({
+  name: r.properties.name,
+  street: r.properties.street,
+  housenumber: r.properties.housenumber,
+  postalcode: r.properties.postalcode,
+  locality: r.properties.locality,
+  coordinates: {
+    lon: r.geometry.coordinates[0],
+    lat: r.geometry.coordinates[1],
+  },
+})
+
+module.exports = {
+  scanAddress,
+  extractAdresses,
+  adress,
+  formatRegexResult,
+  peliasResToSuggestedRecipient,
+}
