+++ packages/booking-interface/wizards/freightslip/handlers.js
@@ -1,9 +1,9 @@
-const { Markup } = require('telegraf')
 const Composer = require('telegraf/composer')
 const bot = require('../../adapters/bot')
 const services = require('../../services')
 const utils = require('../../utils')
 const wizardHelpers = require('../helpers')
+const messaging = require('./messaging')
 
 const awaitManualSenderInput = new Composer().on('text', (ctx) =>
   services.geolocation.get(ctx.update.message.text).then((res) => {
@@ -12,18 +12,7 @@ const awaitManualSenderInput = new Composer().on('text', (ctx) =>
     }
 
     const { state } = ctx.scene.session
-
-    state.suggestedSenders = res.map((r) => ({
-      name: r.properties.name,
-      street: r.properties.street,
-      housenumber: r.properties.housenumber,
-      postalcode: r.properties.postalcode,
-      locality: r.properties.locality,
-      coordinates: {
-        lon: r.geometry.coordinates[0],
-        lat: r.geometry.coordinates[1],
-      },
-    }))
+    state.suggestedSenders = res.map(utils.peliasResToSuggestedRecipient)
 
     return wizardHelpers.jumpToStep(ctx, 'askIfCorrectSuggestedSender')
   })
@@ -36,18 +25,7 @@ const awaitManualRecipientInput = new Composer().on('text', (ctx) =>
     }
 
     const { state } = ctx.scene.session
-
-    state.suggestedRecipients = res.map((r) => ({
-      name: r.properties.name,
-      street: r.properties.street,
-      housenumber: r.properties.housenumber,
-      postalcode: r.properties.postalcode,
-      locality: r.properties.locality,
-      coordinates: {
-        lon: r.geometry.coordinates[0],
-        lat: r.geometry.coordinates[1],
-      },
-    }))
+    state.suggestedRecipients = res.map(utils.peliasResToSuggestedRecipient)
 
     return wizardHelpers.jumpToStep(ctx, 'askIfCorrectSuggestedRecipient')
   })
@@ -113,7 +91,7 @@ const awaitAdditionalInformationOrConfirm = new Composer()
   .action('booking:confirm', (ctx) => {
     ctx.scene.session.state = {}
 
-    return wizardHelpers.jumpToStep(ctx, 'intro')
+    return wizardHelpers.jumpToStep(ctx, 'greet')
   })
   .action('booking:add_extra', (ctx) => {
     console.log('booking wants extra')
@@ -182,18 +160,7 @@ const awaitSenderLocationConfirm = new Composer()
         return wizardHelpers.jumpToStep(ctx, 'notifyNoGeolocationResult')
 
       const { state } = ctx.scene.session
-
-      state.suggestedSenders = res.map((r) => ({
-        name: r.properties.name,
-        street: r.properties.street,
-        housenumber: r.properties.housenumber,
-        postalcode: r.properties.postalcode,
-        locality: r.properties.locality,
-        coordinates: {
-          lon: r.geometry.coordinates[0],
-          lat: r.geometry.coordinates[1],
-        },
-      }))
+      state.suggestedSenders = res.map(utils.peliasResToSuggestedRecipient)
 
       return wizardHelpers.jumpToStep(ctx, 'askIfCorrectSuggestedSender')
     })
@@ -202,42 +169,18 @@ const awaitSenderLocationConfirm = new Composer()
     wizardHelpers.jumpToStep(ctx, 'askForManualRecipient')
   )
 
-const informNoSuggestedSenders = (ctx) => {
-  ctx
-    .reply('Det finns inga förslagna adresser.')
+const informNoSuggestedSenders = (ctx) =>
+  messaging
+    .informNoSuggestedSenders(ctx)
     .then(() => wizardHelpers.jumpToStep(ctx, 'askForSenderLocation'))
-}
-const informNoSuggestedRecipients = (ctx) => {
-  ctx
-    .reply('Det finns inga förslagna adresser.')
+
+const informNoSuggestedRecipients = (ctx) =>
+  messaging
+    .informNoSuggestedRecipients(ctx)
     .then(() => wizardHelpers.jumpToStep(ctx, 'askForManualRecipient'))
-}
 
-const askIfCorrectSuggestedSender = (ctx) => {
-  const { state } = ctx.scene.session
-
-  if (!state.suggestedSenders)
-    return ctx.reply('Vi kunde inte hitta några föreslagna adresser :(')
-
-  const [suggestion] = state.suggestedSenders
-
-  return ctx
-    .replyWithMarkdown(
-      `Är detta rätt?\n\n`
-        .concat(
-          suggestion.street && suggestion.housenumber
-            ? `${suggestion.street} ${suggestion.housenumber}`
-            : suggestion.name
-        )
-        .concat(suggestion.postalcode ? `\n${suggestion.postalcode}` : '')
-        .concat(suggestion.locality ? `\n${suggestion.locality}` : ''),
-      Markup.inlineKeyboard([
-        Markup.callbackButton('Ja', 'sender:geolookup:confirm'),
-        Markup.callbackButton('Nej', 'sender:geolookup:decline'),
-      ]).extra()
-    )
-    .then(() => ctx.wizard.next())
-}
+const askIfCorrectSuggestedSender = (ctx) =>
+  messaging.askIfCorrectSuggestedSender(ctx).then(() => ctx.wizard.next())
 
 const askIfCorrectSuggestedRecipient = (ctx) => {
   const { state } = ctx.scene.session
@@ -245,101 +188,34 @@ const askIfCorrectSuggestedRecipient = (ctx) => {
   if (!state.suggestedRecipients)
     return ctx.reply('Vi kunde inte hitta några föreslagna adresser :(')
 
-  const [suggestion] = state.suggestedRecipients
-
-  return ctx
-    .replyWithMarkdown(
-      `Är detta rätt?\n\n`
-        .concat(
-          suggestion.street && suggestion.housenumber
-            ? `${suggestion.street} ${suggestion.housenumber}`
-            : suggestion.name
-        )
-        .concat(suggestion.postalcode ? `\n${suggestion.postalcode}` : '')
-        .concat(suggestion.locality ? `\n${suggestion.locality}` : ''),
-      Markup.inlineKeyboard([
-        Markup.callbackButton('Ja', 'recipient:geolookup:confirm'),
-        Markup.callbackButton('Nej', 'recipient:geolookup:decline'),
-      ]).extra()
-    )
+  return messaging
+    .askIfCorrectSuggestedRecipient(ctx)
     .then(() => ctx.wizard.next())
 }
 
 const notifyNoGeolocationResult = (ctx) =>
-  ctx
-    .reply(`Vi fick ingen träff på denna adress och detta namn...`)
-    .then(() => wizardHelpers.jumpToStep(ctx, 'intro'))
+  messaging
+    .notifyNoGeolocationResult(ctx)
+    .then(() => wizardHelpers.jumpToStep(ctx, 'greet'))
 
 const askForManualRecipient = (ctx) =>
-  ctx.reply('Ange mottagaradressen').then(() => ctx.wizard.next())
+  messaging.askForManualRecipient(ctx).then(() => ctx.wizard.next())
 
 const askForManualSender = (ctx) =>
-  ctx.reply('Ange avsändaradressen').then(() => ctx.wizard.next())
+  messaging.askForManualSender(ctx).then(() => ctx.wizard.next())
 
-const askAddAdditionalInformation = (ctx) => {
-  const booking = ctx.scene.session.state.booking
-
-  return ctx
-    .replyWithMarkdown(
-      `Då har du fått bokningsnummer:\n\n${ctx.scene.session.state.booking.id}\n\nAnteckna detta på försändelsen.`
-        .concat(`\nSå här ser din bokning ut:`)
-        .concat(`\n\nFrån:\n`)
-        .concat(
-          booking.from.street && booking.from.housenumber
-            ? `${booking.from.street} ${booking.from.housenumber}`
-            : booking.from.name
-        )
-
-        .concat(booking.from.postalcode ? `\n${booking.from.postalcode}` : '')
-        .concat(booking.from.locality ? `\n${booking.from.locality}` : '')
-        .concat(`\n\nTill:\n`)
-        .concat(
-          booking.to.street && booking.to.housenumber
-            ? `${booking.to.street} ${booking.to.housenumber}`
-            : booking.to.name
-        )
-        .concat(booking.to.postalcode ? `\n${booking.to.postalcode}` : '')
-        .concat(booking.to.locality ? `\n${booking.to.locality}` : ''),
-      Markup.inlineKeyboard([
-        Markup.callbackButton('Fyll i fler detaljer', 'booking:add_extra'),
-        Markup.callbackButton('Påbörja nästa', 'booking:confirm'),
-      ]).extra()
-    )
-    .then(() => ctx.wizard.next())
-}
-
-const askForSenderLocation = (ctx) => {
-  return ctx
-    .replyWithMarkdown(
-      `Hur vill du ange avsändaradressen?`,
-      Markup.inlineKeyboard([
-        Markup.callbackButton('Dela position', 'location:from_location'),
-        Markup.callbackButton('Manuellt', 'location:from_manual'),
-      ]).extra()
-    )
-    .then(() => ctx.wizard.next())
-}
-
-const intro = (ctx) =>
-  ctx
-    .replyWithMarkdown(
-      `Har din försändelse en fraktsedel?`,
-      Markup.inlineKeyboard([
-        Markup.callbackButton('Ja', 'freightslip:confirm'),
-        Markup.callbackButton('Nej', 'freightslip:decline'),
-      ]).extra()
-    )
+const askAddAdditionalInformation = (ctx) =>
+  messaging
+    .askAddAdditionalInformation(ctx, ctx.scene.session.state.booking)
     .then(() => ctx.wizard.next())
 
-const askForUpload = (ctx) => {
-  return ctx
-    .reply(
-      'Ta en bild på fraktsedeln eller addresslappen och skicka den till mig!'.concat(
-        `\nFörsök ta bilden så nära det går och i så bra ljus som möjligt.`
-      )
-    )
-    .then(() => ctx.wizard.next())
-}
+const askForSenderLocation = (ctx) =>
+  messaging.askForSenderLocation(ctx).then(() => ctx.wizard.next())
+
+const greet = (ctx) => messaging.greet(ctx).then(() => ctx.wizard.next())
+
+const askForUpload = (ctx) =>
+  messaging.askForUpload(ctx).then(() => ctx.wizard.next())
 
 const awaitImageUpload = new Composer().on('photo', async (ctx) => {
   const photos = ctx.update.message.photo
@@ -354,19 +230,14 @@ const awaitImageUpload = new Composer().on('photo', async (ctx) => {
       return wizardHelpers.jumpToStep(ctx, 'noParseTextFromImageResult')
     }
 
-    const regexResult = Array.from(text.matchAll(utils.adress))
-      .map((res) => res.groups)
-      .filter((group) => group.street && group.nr && group.zipcode)
-      .map((x) =>
-        // do something better here
-        Object.assign({}, x, { zipcode: x.zipcode.replace('SE-', '') })
-      )
+    const regexResult = utils.formatRegexResult(
+      Array.from(text.matchAll(utils.adress))
+    )
 
     const elasticRes = await Promise.all(regexResult.map(services.elastic.get))
-
-    const searchResults = elasticRes
-      .map((res) => res.body)
-      .map(services.formatQueryResult)
+    const searchResults = elasticRes.map((res) =>
+      services.formatQueryResult(res.body)
+    )
 
     if (!searchResults.length) {
       return wizardHelpers.jumpToStep(ctx, 'noParseTextFromImageResult')
@@ -382,55 +253,19 @@ const awaitImageUpload = new Composer().on('photo', async (ctx) => {
   }
 })
 
-const noParseTextFromImageResult = (ctx) => {
-  return ctx
-    .replyWithMarkdown(
-      `Vi kunde inte tolka bilden. Vill du försöka igen?`,
-      Markup.inlineKeyboard([
-        Markup.callbackButton('Ja', 'retry_upload'),
-        Markup.callbackButton('Nej', 'enter_manual'),
-      ]).extra()
-    )
+const noParseTextFromImageResult = (ctx) =>
+  messaging.noParseTextFromImageResult(ctx).then(() => ctx.wizard.next())
+
+const askForSenderOrRecipientConfirmation = (ctx) =>
+  messaging
+    .askForSenderOrRecipientConfirmation(ctx)
     .then(() => ctx.wizard.next())
-}
 
-const askForSenderOrRecipientConfirmation = (ctx) => {
-  const [match] = ctx.scene.session.state.matches || []
-
-  if (!match) return // enter manually or something
-
-  return ctx
-    .replyWithMarkdown(
-      `Så här tolkade vi bilden:`
-        .concat(`\n\n${match.address.street} ${match.address.number}`)
-        .concat(`\n${match.address.zip}`)
-        .concat(`\n${match.locality}`)
-        .concat(`\n\nÄr detta avsändare eller mottagare?`),
-      Markup.inlineKeyboard(
-        [
-          Markup.callbackButton('Mottagare', 'freightslip:is_recipient'),
-          Markup.callbackButton('Avsändare', 'freightslip:is_sender'),
-          Markup.callbackButton('Ingetdera', 'freightslip:is_neither'),
-        ],
-        { resize_keyboard: true }
-      ).extra()
-    )
-    .then(() => ctx.wizard.next())
-}
-
-const askForSenderLocationConfirm = (ctx) => {
-  return ctx
-    .reply('Klicka på knappen för att dela position.', {
-      reply_markup: Markup.keyboard([
-        Markup.locationRequestButton('📍 Dela position'),
-        Markup.callbackButton('Avbryt', 'location:cancel'),
-      ]).oneTime(),
-    })
-    .then(() => ctx.wizard.next())
-}
+const askForSenderLocationConfirm = (ctx) =>
+  messaging.askForSenderLocationConfirm(ctx).then(() => ctx.wizard.next())
 
 module.exports = [
-  intro,
+  greet,
   awaitHasFreightslip,
   askForUpload,
   awaitImageUpload,
@@ -457,4 +292,5 @@ module.exports = [
   informNoSuggestedSenders,
   askIfCorrectSuggestedRecipient,
   awaitManualContactConfirmation,
+  informNoSuggestedRecipients,
 ]
