+++ packages/booking-interface/wizards/freightslip/messaging.js
@@ -1,0 +1,173 @@
+const { Markup } = require('telegraf')
+
+const greet = (ctx) =>
+  ctx.replyWithMarkdown(
+    `Har din försändelse en fraktsedel?`,
+    Markup.inlineKeyboard([
+      Markup.callbackButton('Ja', 'freightslip:confirm'),
+      Markup.callbackButton('Nej', 'freightslip:decline'),
+    ]).extra()
+  )
+
+const askForUpload = (ctx) =>
+  ctx.reply(
+    'Ta en bild på fraktsedeln eller addresslappen och skicka den till mig!'.concat(
+      `\nFörsök ta bilden så nära det går och i så bra ljus som möjligt.`
+    )
+  )
+
+const askForSenderLocation = (ctx) =>
+  ctx.replyWithMarkdown(
+    `Hur vill du ange avsändaradressen?`,
+    Markup.inlineKeyboard([
+      Markup.callbackButton('Dela position', 'location:from_location'),
+      Markup.callbackButton('Manuellt', 'location:from_manual'),
+    ]).extra()
+  )
+
+const askAddAdditionalInformation = (ctx, booking) => {
+  return ctx.replyWithMarkdown(
+    `Då har du fått bokningsnummer:\n\n${ctx.scene.session.state.booking.id}\n\nAnteckna detta på försändelsen.`
+      .concat(`\nSå här ser din bokning ut:`)
+      .concat(`\n\nFrån:\n`)
+      .concat(
+        booking.from.street && booking.from.housenumber
+          ? `${booking.from.street} ${booking.from.housenumber}`
+          : booking.from.name
+      )
+
+      .concat(booking.from.postalcode ? `\n${booking.from.postalcode}` : '')
+      .concat(booking.from.locality ? `\n${booking.from.locality}` : '')
+      .concat(`\n\nTill:\n`)
+      .concat(
+        booking.to.street && booking.to.housenumber
+          ? `${booking.to.street} ${booking.to.housenumber}`
+          : booking.to.name
+      )
+      .concat(booking.to.postalcode ? `\n${booking.to.postalcode}` : '')
+      .concat(booking.to.locality ? `\n${booking.to.locality}` : ''),
+    Markup.inlineKeyboard([
+      Markup.callbackButton('Fyll i fler detaljer', 'booking:add_extra'),
+      Markup.callbackButton('Påbörja nästa', 'booking:confirm'),
+    ]).extra()
+  )
+}
+
+const informNoSuggestedSenders = (ctx) => {
+  ctx.reply('Det finns inga förslagna adresser.')
+}
+
+const informNoSuggestedRecipients = (ctx) => {
+  ctx.reply('Det finns inga förslagna adresser.')
+}
+
+const askIfCorrectSuggestedSender = (ctx) => {
+  const { state } = ctx.scene.session
+
+  if (!state.suggestedSenders)
+    return ctx.reply('Vi kunde inte hitta några föreslagna adresser :(')
+
+  const [suggestion] = state.suggestedSenders
+
+  return ctx.replyWithMarkdown(
+    `Är detta rätt?\n\n`
+      .concat(
+        suggestion.street && suggestion.housenumber
+          ? `${suggestion.street} ${suggestion.housenumber}`
+          : suggestion.name
+      )
+      .concat(suggestion.postalcode ? `\n${suggestion.postalcode}` : '')
+      .concat(suggestion.locality ? `\n${suggestion.locality}` : ''),
+    Markup.inlineKeyboard([
+      Markup.callbackButton('Ja', 'sender:geolookup:confirm'),
+      Markup.callbackButton('Nej', 'sender:geolookup:decline'),
+    ]).extra()
+  )
+}
+
+const askIfCorrectSuggestedRecipient = (ctx) => {
+  const { state } = ctx.scene.session
+
+  if (!state.suggestedRecipients)
+    return ctx.reply('Vi kunde inte hitta några föreslagna adresser :(')
+
+  const [suggestion] = state.suggestedRecipients
+
+  return ctx.replyWithMarkdown(
+    `Är detta rätt?\n\n`
+      .concat(
+        suggestion.street && suggestion.housenumber
+          ? `${suggestion.street} ${suggestion.housenumber}`
+          : suggestion.name
+      )
+      .concat(suggestion.postalcode ? `\n${suggestion.postalcode}` : '')
+      .concat(suggestion.locality ? `\n${suggestion.locality}` : ''),
+    Markup.inlineKeyboard([
+      Markup.callbackButton('Ja', 'recipient:geolookup:confirm'),
+      Markup.callbackButton('Nej', 'recipient:geolookup:decline'),
+    ]).extra()
+  )
+}
+
+const notifyNoGeolocationResult = (ctx) =>
+  ctx.reply(`Vi fick ingen träff på denna adress och detta namn...`)
+
+const askForManualRecipient = (ctx) => ctx.reply('Ange mottagaradressen')
+
+const askForManualSender = (ctx) => ctx.reply('Ange avsändaradressen')
+
+const noParseTextFromImageResult = (ctx) =>
+  ctx.replyWithMarkdown(
+    `Vi kunde inte tolka bilden. Vill du försöka igen?`,
+    Markup.inlineKeyboard([
+      Markup.callbackButton('Ja', 'retry_upload'),
+      Markup.callbackButton('Nej', 'enter_manual'),
+    ]).extra()
+  )
+
+const askForSenderOrRecipientConfirmation = (ctx) => {
+  const [match] = ctx.scene.session.state.matches || []
+
+  if (!match) return // enter manually or something
+
+  return ctx.replyWithMarkdown(
+    `Så här tolkade vi bilden:`
+      .concat(`\n\n${match.address.street} ${match.address.number}`)
+      .concat(`\n${match.address.zip}`)
+      .concat(`\n${match.locality}`)
+      .concat(`\n\nÄr detta avsändare eller mottagare?`),
+    Markup.inlineKeyboard(
+      [
+        Markup.callbackButton('Mottagare', 'freightslip:is_recipient'),
+        Markup.callbackButton('Avsändare', 'freightslip:is_sender'),
+        Markup.callbackButton('Ingetdera', 'freightslip:is_neither'),
+      ],
+      { resize_keyboard: true }
+    ).extra()
+  )
+}
+
+const askForSenderLocationConfirm = (ctx) =>
+  ctx.reply('Klicka på knappen för att dela position.', {
+    reply_markup: Markup.keyboard([
+      Markup.locationRequestButton('📍 Dela position'),
+      Markup.callbackButton('Avbryt', 'location:cancel'),
+    ]).oneTime(),
+  })
+
+module.exports = {
+  greet,
+  noParseTextFromImageResult,
+  askForUpload,
+  askForSenderOrRecipientConfirmation,
+  askForSenderLocationConfirm,
+  askForSenderLocation,
+  askAddAdditionalInformation,
+  informNoSuggestedSenders,
+  informNoSuggestedRecipients,
+  askIfCorrectSuggestedSender,
+  askIfCorrectSuggestedRecipient,
+  notifyNoGeolocationResult,
+  askForManualRecipient,
+  askForManualSender,
+}
